--- code/cell.electron/app/src/main/main.sys/client.ts
@@ -1,5 +1,0 @@
-import '../config';
-
-export * from './client.sys';
-export * from './client.typeDefs';
-export * from './client.upload';
