+++ code/cell.electron/app/src/main/main.sys/index.ts
@@ -1,2 +1,2 @@
-import * as sys from './client';
+import * as sys from './sys';
 export { sys };
