+++ code/cell.electron/app/src/main/main.sys/sys.ts
@@ -1,0 +1,5 @@
+import '../config';
+
+export * from './sys.init';
+export * from './sys.typeDefs';
+export * from './sys.upload';
